video: tegra: dc: improve underflow statistics
Collect individual underflow counts for underflow statistics to provide
a more accurate number of underflows. Changed stats to use 64-bit numbers
due to the larger numbers involved, about 100x from previously.

Reviewed-on: http://git-master/r/52940
(cherry picked from commit 1d39d430ad90a027be43323f65eef85a3a30faab)

Reviewed-on: http://git-master/r/56301
Tested-by: Varun Colbert <vcolbert@nvidia.com>

+++ drivers/video/tegra/dc/dc_priv.h
@@ -126,10 +126,10 @@ struct tegra_dc {
 	struct work_struct		vblank_work;
 
 	struct {
-		unsigned		underflows;
-		unsigned		underflows_a;
-		unsigned		underflows_b;
-		unsigned		underflows_c;
+		u64			underflows;
+		u64			underflows_a;
+		u64			underflows_b;
+		u64			underflows_c;
 	} stats;
 
 	struct tegra_dc_ext		*ext;
